general enhancements interals
