+++ src/Foundation/Attributes/RequireLivewireContext.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace TallStackUi\Foundation\Attributes;
+
+use Attribute;
+
+#[Attribute(Attribute::TARGET_CLASS)]
+class RequireLivewireContext
+{
+    //
+}
