+++ src/Foundation/Exceptions/MissingLivewireException.php
@@ -10,24 +10,4 @@ class MissingLivewireException extends Exception
     {
         parent::__construct('The ['.$component.'] component should only be used inside Livewire components.');
     }
-
-    /** @throws MissingLivewireException */
-    public static function throwIf(bool $boolean, string $component): ?self
-    {
-        if ($boolean) {
-            return null;
-        }
-
-        throw new self($component);
-    }
-
-    /** @throws MissingLivewireException */
-    public static function throwUnless(bool $boolean, string $component): ?self
-    {
-        if (! $boolean) {
-            return null;
-        }
-
-        throw new self($component);
-    }
 }
