+++ src/Foundation/Support/Blade/BladeComponentPrefix.php
@@ -9,12 +9,21 @@ class BladeComponentPrefix
         $this->prefix = config('tallstackui.prefix');
     }
 
-    public function __invoke(string $component): string
+    public function add(string $component): string
     {
         if (blank($this->prefix) || (bool) $this->prefix === false) {
             return $component;
         }
 
         return str($component)->start($this->prefix)->value();
+    }
+
+    public function remove(string $component): string
+    {
+        if (blank($this->prefix) || (bool) $this->prefix === false) {
+            return $component;
+        }
+
+        return str($component)->remove($this->prefix)->value();
     }
 }
