+++ src/View/Components/Progress/Traits/SetupProgress.php
@@ -4,7 +4,7 @@ namespace TallStackUi\View\Components\Progress\Traits;
 
 trait SetupProgress
 {
-    public function setup(): void
+    private function setup(): void
     {
         $this->size = $this->xs ? 'xs' : ($this->sm ? 'sm' : ($this->lg ? 'lg' : 'md'));
