+++ src/resources/views/components/loading.blade.php
@@ -1,5 +1,4 @@
 @php
-    \TallStackUi\Foundation\Exceptions\MissingLivewireException::throwIf($livewire, 'loading');
     $personalize = $classes();
 @endphp
