+++ src/resources/views/components/reaction.blade.php
@@ -1,5 +1,4 @@
 @php
-    \TallStackUi\Foundation\Exceptions\MissingLivewireException::throwIf($livewire, 'reaction');
     $entangle = $bind($attributes, livewire: $livewire)[3];
     $personalize = $classes();
     $extension = $animated === true ? 'gif' : 'png';
